latch file api

+++ modules/nf-commons/src/main/nextflow/file/CopyMoveHelper.java
@@ -87,7 +87,7 @@ public class CopyMoveHelper {
         }
 
         try (InputStream in = Files.newInputStream(source)) {
-            Files.copy(in, target);
+            Files.copy(in, target, StandardCopyOption.REPLACE_EXISTING);
         }
     }
 
@@ -102,7 +102,6 @@ public class CopyMoveHelper {
     static void copyDirectory( final Path source, final Path target, final CopyOption... options )
             throws IOException
     {
-
         final boolean foreign = source.getFileSystem().provider() != target.getFileSystem().provider();
 
         FileVisitor<Path> visitor = new SimpleFileVisitor<Path>() {
@@ -157,11 +156,13 @@ public class CopyMoveHelper {
             throw new IOException("Copying of symbolic links not supported");
 
         // delete target if it exists and REPLACE_EXISTING is specified
-        if (opts.replaceExisting()) {
-            FileHelper.deletePath(target);
+        if (!target.getFileSystem().provider().getScheme().equals("latch")) {
+            if (opts.replaceExisting()) {
+                FileHelper.deletePath(target);
+            } else if (Files.exists(target)) {
+                throw new FileAlreadyExistsException(target.toString());
+            }
         }
-        else if (Files.exists(target))
-            throw new FileAlreadyExistsException(target.toString());
 
         // create directory or copy file
         if (attrs.isDirectory()) {
